Ingress: match all hosts by omitting `host`
Previously, a `host` was mandatory, and `*` is not supported by
kubernetes as it is an invalid domain name.

In kubernetes Ingress when the `host` is not specified in a rule, then
it applies to all hosts of incoming requests.

An alternate implementation could be to explicitly expose to helm
values the `defaultBackend`.