Allow the route termination to be customized
This covers, in part, issue #490. Namely, the route currently hardcodes
the termination mode to "passhtrough". Let's parametrize this so a user
can customize it. The use-case here being that we can deploy the vault
in http mode and expose it via https externally and by choosing 'edge'
as the termination type we let the TLS be taken care of by the route
object.

The default remains 'passthrough' in order to not disrupt any existing
setup.