fix(test): add securityContext for Helm test
When running tests with Helm the same securityContext also should apply.
Therefore this change will add the securityContext from the values file also to the Helm tests.

The change will use the securityContext of the server statefulSet.
This results from the fact that the same image is used in the test.

Signed-off-by: Marco Lecheler <marco.lecheler@mercedes-benz.com>